Tidy up and add (AI generated) tests

+++ 04-json-parser/src/number_parser.zig
@@ -21,8 +21,8 @@ pub const NumberParser = struct {
         return c;
     }
 
-    pub fn parse(self: *Self) f32 {
-        const sign_multiplier: f32 = if (self.peek().? == '-') -1.0 else 1.0;
+    pub fn parse(self: *Self) f64 {
+        const sign_multiplier: f64 = if (self.peek().? == '-') -1.0 else 1.0;
         if (sign_multiplier == -1.0) {
             _ = self.consume();
         }
@@ -33,11 +33,11 @@ pub const NumberParser = struct {
         // Parse exponent part (if present)
         const exponent_part = self.parseExponentPart();
         // Bring it all together
-        return sign_multiplier * ((whole_part + fraction_part) * std.math.pow(f32, 10, exponent_part));
+        return sign_multiplier * ((whole_part + fraction_part) * std.math.pow(f64, 10, exponent_part));
     }
 
-    fn parseWholePart(self: *Self) f32 {
-        var whole_part: f32 = 0.0;
+    fn parseWholePart(self: *Self) f64 {
+        var whole_part: f64 = 0.0;
         // Validated number is guaranteed to have some whole part
         const is_zero = (self.peek().?) == '0';
         if (is_zero) {
@@ -46,7 +46,7 @@ pub const NumberParser = struct {
             while (self.peek()) |char| {
                 if (!std.ascii.isDigit(char)) break;
                 _ = self.consume();
-                const value: f32 = @floatFromInt(char - 48);
+                const value: f64 = @floatFromInt(char - 48);
                 whole_part *= 10.0;
                 whole_part += value;
             }
@@ -54,34 +54,34 @@ pub const NumberParser = struct {
         return whole_part;
     }
 
-    fn parseFractionPart(self: *Self) f32 {
-        var fraction_part: f32 = 0.0;
+    fn parseFractionPart(self: *Self) f64 {
+        var fraction_part: f64 = 0.0;
         const contains_fraction_part = if (self.peek()) |char| char == '.' else false;
         if (!contains_fraction_part) return fraction_part;
 
         // Consume the decimal
         _ = self.consume();
         // Parse the digits
-        var coefficient: f32 = 0.1;
+        var coefficient: f64 = 0.1;
         while (self.peek()) |char| {
             if (!std.ascii.isDigit(char)) break;
             _ = self.consume();
-            const value: f32 = @floatFromInt(char - 48);
+            const value: f64 = @floatFromInt(char - 48);
             fraction_part += (value * coefficient);
             coefficient *= 0.1;
         }
         return fraction_part;
     }
 
-    fn parseExponentPart(self: *Self) f32 {
-        var exponent_part: f32 = 0.0;
+    fn parseExponentPart(self: *Self) f64 {
+        var exponent_part: f64 = 0.0;
         const contains_exponent_part = if (self.peek()) |char| (std.ascii.toLower(char) == 'e') else false;
         if (!contains_exponent_part) return exponent_part;
 
         // Consume the exponent symbol
         _ = self.consume();
         // There can be a sign after the exponent part
-        const exponent_sign_multiplier: f32 = if ((self.peek().?) == '-') -1.0 else 1.0;
+        const exponent_sign_multiplier: f64 = if ((self.peek().?) == '-') -1.0 else 1.0;
         if (self.peek()) |char| {
             if (char == '-' or char == '+') _ = self.consume();
         }
@@ -90,7 +90,7 @@ pub const NumberParser = struct {
         while (self.peek()) |char| {
             if (!std.ascii.isDigit(char)) break;
             _ = self.consume();
-            const value: f32 = @floatFromInt(char - 48);
+            const value: f64 = @floatFromInt(char - 48);
             exponent_part *= 10.0;
             exponent_part += value;
         }
